Renaming Game Type Enums. Tidying up package. Adding new Difficulty enum.
