+++ gazeplay-games/src/main/java/net/gazeplay/games/whereisit/Difficulty.java
@@ -1,0 +1,5 @@
+package net.gazeplay.games.whereisit;
+
+public enum Difficulty {
+    EASY, NORMAL, HARD
+}
