+++ gazeplay-games/src/main/java/net/gazeplay/games/whereisit/WhereIsItGameType.java
@@ -3,11 +3,14 @@ package net.gazeplay.games.whereisit;
 import lombok.Getter;
 
 public enum WhereIsItGameType {
-    ANIMALNAME("where-is-the-animal", "where-is-the-animal"), COLORNAME("where-is-the-color",
-        "where-is-the-color"), COLORNAMEEASY("where-is-the-color",
-        "where-is-the-color-easy"), LETTERS("where-is-the-letter", "where-is-the-letter"), NUMBERS(
-        "where-is-the-number", "where-is-the-number"), FLAGS("find-flag", "find-flag"), CUSTOMIZED(
-        "customized", "customized"), FINDODD("find-the-odd-one-out", "find-the-odd-one-out");
+    ANIMAL_NAME("where-is-the-animal", "where-is-the-animal"),
+    COLOR_NAME("where-is-the-color", "where-is-the-color"),
+    COLOR_NAME_EASY("where-is-the-color", "where-is-the-color", Difficulty.EASY),
+    LETTERS("where-is-the-letter", "where-is-the-letter"),
+    NUMBERS("where-is-the-number", "where-is-the-number"),
+    FLAGS("find-flag", "find-flag"),
+    CUSTOMIZED("customized", "customized"),
+    FIND_ODD("find-the-odd-one-out", "find-the-odd-one-out");
 
     @Getter
     private final String gameName;
@@ -18,9 +21,17 @@ public enum WhereIsItGameType {
     @Getter
     private final String languageResourceLocation;
 
+    @Getter
+    private final Difficulty difficulty;
+
     WhereIsItGameType(String gameName, String resourcesDirectoryName) {
+        this(gameName, resourcesDirectoryName, Difficulty.NORMAL);
+    }
+
+    WhereIsItGameType(String gameName, String resourcesDirectoryName, Difficulty difficulty) {
         this.gameName = gameName;
         this.resourcesDirectoryName = resourcesDirectoryName;
         this.languageResourceLocation = "data/" + resourcesDirectoryName + "/" + resourcesDirectoryName + ".csv";
+        this.difficulty = difficulty;
     }
 }
