+++ gazeplay-games/src/main/java/net/gazeplay/games/whereisit/launcher/FlagsGameLauncher.java
@@ -1,10 +1,13 @@
-package net.gazeplay.games.whereisit;
+package net.gazeplay.games.whereisit.launcher;
 
 import javafx.scene.Scene;
 import net.gazeplay.GameLifeCycle;
 import net.gazeplay.GameSpec;
 import net.gazeplay.IGameContext;
 import net.gazeplay.commons.utils.stats.Stats;
+import net.gazeplay.games.whereisit.WhereIsIt;
+import net.gazeplay.games.whereisit.WhereIsItGameType;
+import net.gazeplay.games.whereisit.WhereIsItStats;
 
 public class FlagsGameLauncher implements GameSpec.GameLauncher {
     @Override
