+++ gazeplay-games/src/main/java/net/gazeplay/games/whereisit/launcher/WhereIsTheLetterGameLauncher.java
@@ -1,21 +1,25 @@
-package net.gazeplay.games.whereisit;
+package net.gazeplay.games.whereisit.launcher;
 
 import javafx.scene.Scene;
 import net.gazeplay.GameLifeCycle;
 import net.gazeplay.GameSpec;
 import net.gazeplay.IGameContext;
 import net.gazeplay.commons.utils.stats.Stats;
+import net.gazeplay.games.whereisit.WhereIsIt;
+import net.gazeplay.games.whereisit.WhereIsItGameType;
+import net.gazeplay.games.whereisit.WhereIsItStats;
 
-public class WhereIsItGameLauncher implements GameSpec.GameLauncher<Stats, GameSpec.DimensionGameVariant> {
+public class WhereIsTheLetterGameLauncher implements GameSpec.GameLauncher<Stats, GameSpec.DimensionGameVariant> {
     @Override
     public Stats createNewStats(Scene scene) {
-        return new WhereIsItStats(scene, WhereIsItGameType.CUSTOMIZED.getGameName());
+        return new WhereIsItStats(scene, WhereIsItGameType.LETTERS.getGameName());
     }
 
     @Override
     public GameLifeCycle createNewGame(IGameContext gameContext,
                                        GameSpec.DimensionGameVariant gameVariant, Stats stats) {
-        return new WhereIsIt(WhereIsItGameType.CUSTOMIZED, gameVariant.getWidth(),
+        return new WhereIsIt(WhereIsItGameType.LETTERS, gameVariant.getWidth(),
             gameVariant.getHeight(), false, gameContext, stats);
     }
+
 }
