+++ gazeplay-games/src/main/java/net/gazeplay/games/whereisit/launcher/WhereIsTheNumberGameLauncher.java
@@ -1,21 +1,25 @@
-package net.gazeplay.games.whereisit;
+package net.gazeplay.games.whereisit.launcher;
 
 import javafx.scene.Scene;
 import net.gazeplay.GameLifeCycle;
 import net.gazeplay.GameSpec;
 import net.gazeplay.IGameContext;
 import net.gazeplay.commons.utils.stats.Stats;
+import net.gazeplay.games.whereisit.WhereIsIt;
+import net.gazeplay.games.whereisit.WhereIsItGameType;
+import net.gazeplay.games.whereisit.WhereIsItStats;
 
-public class FindOddGameLauncher implements GameSpec.GameLauncher<Stats, GameSpec.DimensionGameVariant> {
+public class WhereIsTheNumberGameLauncher implements GameSpec.GameLauncher<Stats, GameSpec.DimensionGameVariant> {
     @Override
     public Stats createNewStats(Scene scene) {
-        return new WhereIsItStats(scene, WhereIsItGameType.FINDODD.getGameName());
+        return new WhereIsItStats(scene, WhereIsItGameType.NUMBERS.getGameName());
     }
 
     @Override
     public GameLifeCycle createNewGame(IGameContext gameContext,
                                        GameSpec.DimensionGameVariant gameVariant, Stats stats) {
-        return new WhereIsIt(WhereIsItGameType.FINDODD, gameVariant.getWidth(),
+        return new WhereIsIt(WhereIsItGameType.NUMBERS, gameVariant.getWidth(),
             gameVariant.getHeight(), false, gameContext, stats);
     }
+
 }
