+++ gazeplay-games/src/main/java/net/gazeplay/games/whereisit/spec/FindOddGameSpecSource.java
@@ -1,18 +1,20 @@
-package net.gazeplay.games.whereisit;
+package net.gazeplay.games.whereisit.spec;
 
 import net.gazeplay.GameCategories;
 import net.gazeplay.GameSpec;
 import net.gazeplay.GameSpecSource;
 import net.gazeplay.GameSummary;
+import net.gazeplay.games.whereisit.WhereIsItGameVariantGenerator;
+import net.gazeplay.games.whereisit.launcher.FindOddGameLauncher;
 
-public class WhereIsTheAnimalGameSpecSource implements GameSpecSource {
+public class FindOddGameSpecSource implements GameSpecSource {
     @Override
     public GameSpec getGameSpec() {
         return new GameSpec(
-            GameSummary.builder().nameCode("WhereIsTheAnimal").gameThumbnail("data/Thumbnails/whereisanimal.png")
+            GameSummary.builder().nameCode("findodd").gameThumbnail("data/Thumbnails/findtheodd.jpg")
                 .category(GameCategories.Category.SELECTION)
                 .category(GameCategories.Category.LITERACY)
                 .category(GameCategories.Category.MEMORIZATION).build(),
-            new WhereIsItGameVariantGenerator(), new WhereIsTheAnimalGameLauncher());
+            new WhereIsItGameVariantGenerator(), new FindOddGameLauncher());
     }
 }
