+++ gazeplay-games/src/main/java/net/gazeplay/games/whereisit/spec/FlagsGameSpecSource.java
@@ -1,9 +1,10 @@
-package net.gazeplay.games.whereisit;
+package net.gazeplay.games.whereisit.spec;
 
 import net.gazeplay.GameCategories;
 import net.gazeplay.GameSpec;
 import net.gazeplay.GameSpecSource;
 import net.gazeplay.GameSummary;
+import net.gazeplay.games.whereisit.launcher.FlagsGameLauncher;
 
 public class FlagsGameSpecSource implements GameSpecSource {
     @Override
