+++ gazeplay-games/src/main/java/net/gazeplay/games/whereisit/spec/WhereIsTheAnimalGameSpecSource.java
@@ -1,0 +1,20 @@
+package net.gazeplay.games.whereisit.spec;
+
+import net.gazeplay.GameCategories;
+import net.gazeplay.GameSpec;
+import net.gazeplay.GameSpecSource;
+import net.gazeplay.GameSummary;
+import net.gazeplay.games.whereisit.WhereIsItGameVariantGenerator;
+import net.gazeplay.games.whereisit.launcher.WhereIsTheAnimalGameLauncher;
+
+public class WhereIsTheAnimalGameSpecSource implements GameSpecSource {
+    @Override
+    public GameSpec getGameSpec() {
+        return new GameSpec(
+            GameSummary.builder().nameCode("WhereIsTheAnimal").gameThumbnail("data/Thumbnails/whereisanimal.png")
+                .category(GameCategories.Category.SELECTION)
+                .category(GameCategories.Category.LITERACY)
+                .category(GameCategories.Category.MEMORIZATION).build(),
+            new WhereIsItGameVariantGenerator(), new WhereIsTheAnimalGameLauncher());
+    }
+}
